textdiection rtl/ltr support added
